requirements.txt
# Core dependencies
pyautogui>=0.9.54
Pillow>=10.0.0
requests>=2.31.0

# Enhanced functionality
hashlib2>=1.0.1
psutil>=5.9.0

# Development and testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Built-in modules (no installation required):
# - json
# - logging
# - time
# - base64
# - typing
# - pathlib
# - argparse
# - sys
# - os
# - abc
# - dataclasses
# - webbrowser
# - threading
# - hashlib
# - io

# Note: Install with: pip install -r requirements.txt
# For development: pip install -r requirements.txt pytest black flake8